+++ pkg/store/proxy.go
@@ -195,14 +195,31 @@ func NewProxyStore(
 }
 
 func (s *ProxyStore) LabelSet() []labelpb.ZLabelSet {
-	stores := s.stores()
+	stores := s.storesForTSDBSelector()
 	if len(stores) == 0 {
+		// We always want to enforce announcing the subset of data that
+		// selector-labels represents. If no stores match the filter,
+		// we still want to enforce announcing this subset.
+		selectorLabels := labelpb.ZLabelsFromPromLabels(s.selectorLabels)
+		if len(selectorLabels) > 0 {
+			return []labelpb.ZLabelSet{{Labels: selectorLabels}}
+		}
 		return []labelpb.ZLabelSet{}
 	}
 
 	mergedLabelSets := make(map[uint64]labelpb.ZLabelSet, len(stores))
 	for _, st := range stores {
-		for _, lset := range st.LabelSets() {
+		// Get filtered label sets from TSDBSelector
+		_, filteredLabelSets := s.tsdbSelector.MatchLabelSets(st.LabelSets()...)
+
+		var labelSetsToProcess []labels.Labels
+		if filteredLabelSets != nil {
+			labelSetsToProcess = filteredLabelSets
+		} else {
+			labelSetsToProcess = st.LabelSets()
+		}
+
+		for _, lset := range labelSetsToProcess {
 			mergedLabelSet := labelpb.ExtendSortedLabels(lset, s.selectorLabels)
 			mergedLabelSets[mergedLabelSet.Hash()] = labelpb.ZLabelSet{Labels: labelpb.ZLabelsFromPromLabels(mergedLabelSet)}
 		}
@@ -226,14 +243,14 @@ func (s *ProxyStore) LabelSet() []labelpb.ZLabelSet {
 }
 
 func (s *ProxyStore) TimeRange() (int64, int64) {
-	stores := s.stores()
+	stores := s.storesForTSDBSelector()
 	if len(stores) == 0 {
 		return math.MinInt64, math.MaxInt64
 	}
 
 	var minTime, maxTime int64 = math.MaxInt64, math.MinInt64
-	for _, s := range stores {
-		storeMinTime, storeMaxTime := s.TimeRange()
+	for _, st := range stores {
+		storeMinTime, storeMaxTime := st.TimeRange()
 		if storeMinTime < minTime {
 			minTime = storeMinTime
 		}
@@ -246,12 +263,9 @@ func (s *ProxyStore) TimeRange() (int64, int64) {
 }
 
 func (s *ProxyStore) TSDBInfos() []infopb.TSDBInfo {
+	stores := s.storesForTSDBSelector()
 	infos := make([]infopb.TSDBInfo, 0)
-	for _, st := range s.stores() {
-		matches, _ := s.tsdbSelector.MatchLabelSets(st.LabelSets()...)
-		if !matches {
-			continue
-		}
+	for _, st := range stores {
 		infos = append(infos, st.TSDBInfos()...)
 	}
 	return infos
@@ -572,6 +586,20 @@ func storeInfo(st Client) (storeID string, storeAddr string, isLocalStore bool) 
 	return storeID, storeAddr, isLocalStore
 }
 
+// storesForTSDBSelector returns stores that match the TSDBSelector filtering criteria.
+// This centralizes the TSDBSelector filtering logic used across all ProxyStore methods
+// for cases where we don't need additional matchers or time range filtering.
+func (s *ProxyStore) storesForTSDBSelector() []Client {
+	var filteredStores []Client
+	for _, st := range s.stores() {
+		matches, _ := s.tsdbSelector.MatchLabelSets(st.LabelSets()...)
+		if matches {
+			filteredStores = append(filteredStores, st)
+		}
+	}
+	return filteredStores
+}
+
 // TODO: consider moving the following functions into something like "pkg/pruneutils" since it is also used for exemplars.
 
 func (s *ProxyStore) matchingStores(ctx context.Context, minTime, maxTime int64, matchers []*labels.Matcher) ([]Client, []labels.Labels, []string) {
@@ -580,7 +608,7 @@ func (s *ProxyStore) matchingStores(ctx context.Context, minTime, maxTime int64,
 		storeLabelSets []labels.Labels
 		storeDebugMsgs []string
 	)
-	for _, st := range s.stores() {
+	for _, st := range s.storesForTSDBSelector() {
 		// We might be able to skip the store if its meta information indicates it cannot have series matching our query.
 		if ok, reason := storeMatches(ctx, s.debugLogging, st, minTime, maxTime, matchers...); !ok {
 			if s.debugLogging {
@@ -588,13 +616,8 @@ func (s *ProxyStore) matchingStores(ctx context.Context, minTime, maxTime int64,
 			}
 			continue
 		}
-		matches, extraMatchers := s.tsdbSelector.MatchLabelSets(st.LabelSets()...)
-		if !matches {
-			if s.debugLogging {
-				storeDebugMsgs = append(storeDebugMsgs, fmt.Sprintf("Store %s filtered out due to: %v", st, "tsdb selector"))
-			}
-			continue
-		}
+		// Since we already filtered by TSDBSelector in filteredStores(), we just need to get the extra matchers
+		_, extraMatchers := s.tsdbSelector.MatchLabelSets(st.LabelSets()...)
 		storeLabelSets = append(storeLabelSets, extraMatchers...)
 
 		stores = append(stores, st)
@@ -669,7 +692,7 @@ func storeMatchDebugMetadata(s Client, debugLogging bool, storeDebugMatchers [][
 	return true, ""
 }
 
-// LabelSetsMatch returns false if all label-set do not match the matchers (aka: OR is between all label-sets).
+// LabelSetsMatch returns false if all label-sets do not match the matchers (aka: OR is between all label-sets).
 func LabelSetsMatch(matchers []*labels.Matcher, lset ...labels.Labels) bool {
 	if len(lset) == 0 {
 		return true